backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
redis==5.0.1
groq==0.4.2
openai==1.6.1
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
python-whois==0.9.4
qrcode[pil]==7.4.2
reportlab==4.0.8
bcrypt==4.1.2
aiosqlite==0.19.0
anthropic==0.42.0
